refactor: flatten src/server subdirectories to server root
