+++ src/server/roleConfigurationInterface.ts
