+++ src/server/server.ts
@@ -6,12 +6,12 @@ import ejs from 'ejs';
 import path from 'path';
 import { Auth, newLoginHandler } from 'blaise-login-react/blaise-login-react-server';
 import SurveyHandler from './handlers/surveyHandler';
-import ConfigurationProvider from './configuration/ServerConfigurationProvider';
-import BlaiseApi from './api/BlaiseApi';
+import ConfigurationProvider from './ServerConfigurationProvider';
+import BlaiseApi from './BlaiseApi';
 import BlaiseApiClient from 'blaise-api-node-client';
 import createLogger from "./pino";
 import { HttpLogger } from "pino-http";
-import AuditLogger from "./logger/auditLogger";
+import AuditLogger from "./auditLogger";
 import CaseHandler from './handlers/caseHandler';
 import newClientLogHandler from "./handlers/clientLogHandler";
 import UserHandler from './handlers/userHandler';
