+++ src/server/serverConfigurationInterface.ts
