+++ src/test/server/controllers/surveyController.test.ts
@@ -4,11 +4,11 @@ import { Auth } from 'blaise-login-react/blaise-login-react-server';
 import BlaiseApiClient, { User } from 'blaise-api-node-client';
 import nodeServer from '../../../server/server';
 import createAxiosError from './axiosTestHelper';
-import BlaiseApi from '../../../server/api/BlaiseApi';
+import BlaiseApi from '../../../server/BlaiseApi';
 import FakeServerConfigurationProvider from '../configuration/FakeServerConfigurationProvider';
 import { QuestionnaireDetails, Survey } from '../../../common/interfaces/surveyInterface';
 import userMockObject from '../mockObjects/userMockObject';
-import AuditLogger from '../../../server/logger/auditLogger';
+import AuditLogger from '../../../server/auditLogger';
 
 // create fake config
 const configFake = new FakeServerConfigurationProvider();
