+++ src/test/server/controllers/userController.test.ts
@@ -3,11 +3,11 @@ import { IMock, It, Mock, Times } from 'typemoq';
 import { Auth } from 'blaise-login-react/blaise-login-react-server';
 import BlaiseApiClient, { User } from 'blaise-api-node-client';
 import nodeServer from '../../../server/server';
-import BlaiseApi from '../../../server/api/BlaiseApi';
+import BlaiseApi from '../../../server/BlaiseApi';
 import FakeServerConfigurationProvider from '../configuration/FakeServerConfigurationProvider';
 import createAxiosError from './axiosTestHelper';
 import userMockObject from '../mockObjects/userMockObject';
-import AuditLogger from '../../../server/logger/auditLogger';
+import AuditLogger from '../../../server/auditLogger';
 
 // create fake config
 const configFake = new FakeServerConfigurationProvider();
