+++ src/test/server/mappers/caseMapper.test.ts
@@ -4,7 +4,7 @@ import {
 import { CaseResponse } from 'blaise-api-node-client';
 import { caseResponseMockObject } from '../mockObjects/CaseMockObject';
 import { CaseSummaryDetails, HousingBenefits } from '../../../common/interfaces/caseInterface';
-import mapCaseSummary from '../../../server/mappers/caseMapper';
+import mapCaseSummary from '../../../server/caseMapper';
 
 function SetFieldsToValue(CaseResponseData: CaseResponse, fieldEndsWith: string, fieldValue: string): void {
   Object.keys(CaseResponseData.fieldData).forEach((key) => {
